core: Move FontLike to font/font_like.rs
This makes font.rs smaller, as it's a pretty big file.
